% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{Creates instance of Condition object}
\usage{
condition(name)
}
\arguments{
\item{name}{Name of condition with type string}
}
\value{
A Condition object
}
\description{
Creates instance of Condition object
}
\examples{
condition = condition("latex")
}
