% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{Experiment}
\alias{Experiment}
\title{Creates Instance of Experiment Object}
\usage{
Experiment()
}
\arguments{
\item{tasks}{List of Task objects, which represent the activities the participants will complete}

\item{conditions}{List of condition objects, which represent a variation of the independent variables being tested}

\item{trials}{List of trial objects, which represent an instance of a task completed by a single participant}
}
\value{
An Experiment object
}
\description{
Creates Instance of Experiment Object
}
\examples{
exp <- Experiment()
}
