% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\name{task}
\alias{task}
\title{Creates instance of Task object}
\usage{
task(name)
}
\arguments{
\item{name}{Name of task with type string}
}
\value{
A Task object
}
\description{
Creates instance of Task object
}
\examples{
task = task("short editing task")
}
